+++ changelog.md
@@ -1,4 +1,8 @@
 # Changelog
+### v1.2.6 (22.07.2021)
+* New examples and Demo using from "mens-shoe-prices" DataSet
+* Fixed findAcceptableFiltersCount logics for cases with empty filters
+* Readme updates
 
 ### v1.2.5 (21.05.2021)
 * findAcceptableFiltersCount method added, return acceptable filters with values count